+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Add ambiguous unit resolution depending on context and user locale. `pound` can now become `GBP` or `lbs`
 - Add currency symbols like `€`
 - Add `as` and `into` keywords for unit conversion
 - Fix some units not being simplified